// src/gatz/types.ts
export type SUUID = string;
export type SULID = string;
export type SDate = string;

export type NotificationSettings = {
  overall: boolean;
  activity: "daily" | "none";
  subscribe_on_comment: boolean;
  suggestions_from_gatz: boolean;

  // comments_to_own_post: boolean;
  // reactions_to_own_post: boolean;
  // replies_to_comment: boolean;
  // reactions_to_comment: boolean;
  // at_mentions: boolean;
};

export type AppManifest = {
  app: {
    min_version: string;
    upgrade_message?: string;
    install_links: { ios: string; android: string };
    blocked_version?: string;
  };
};

export type UserProfile = {
  full_name: string | null;
  urls?: {
    website: string | null;
    twitter: string | null;
  };
};

export type LocationSettings = {
  enabled: boolean;
};

export type UserSettings = {
  notifications: NotificationSettings;
  location: LocationSettings;
};

export type Location = {
  id: SUUID,
  name: string,
}

export type NewLocationResponse = {
  location: Location,
  in_common: {
    friends: Contact["id"][],
    friends_of_friends: Contact["id"][],
  }
}

export type User = {
  id: SUUID;
  name: string;
  clock: HLC;
  avatar: string;
  phone_number: string;
  created_at: SDate;
  updated_at: SDate;
  is_test?: boolean;
  is_admin?: boolean;
  push_tokens?: { expo: PushToken };
  settings?: UserSettings;
  profile?: UserProfile;
};

type PushToken = {
  type: "expo";
  created_at: SDate;
  push_token: string;
};

export type HLC = {
  counter: number;
  node: SUUID;
  ts: SDate;
};

export interface CRDT {
  id: string;
  clock: HLC;
}

export type Mention = {
  id: SULID;
  to_uid: User["id"];
  by_uid: User["id"];
  mid: Message["id"];
  did: Discussion["id"];
  ts: SDate;
};

type MemberMode = "open" | "closed" | "friends_of_friends";

export const isOpen = (member_mode: MemberMode) =>
  member_mode === "open" || member_mode === "friends_of_friends";

export type Discussion = {
  id: SUUID;
  clock: HLC;
  muted: boolean;
  type: "discussion";
  name?: string;
  created_by: User["id"];
  group_id?: Group["id"];
  created_at: SDate;
  updated_at: SDate;
  seen_at?: Record<User["id"], string>;
  archived_uids: User["id"][];
  last_message_read?: Record<User["id"], Message["id"]>;
  first_message: Message["id"];
  latest_message: Message["id"];
  latest_activity_ts: SDate;
  members: User["id"][];
  subscribers: User["id"][];
  active_members: User["id"][];
  originally_from?: { did: Discussion["id"]; mid: Message["id"] };
  mentions?: Record<User["id"], Mention[]>;
  member_mode: MemberMode;
  public_mode: "public" | "hidden";
  open_until?: SDate;
  location_id?: Location["id"];
  location?: Location;
};

export type MessageEdits = {
  text: string;
  edited_at: SDate;
};

type ReactionStr = string;

export type Message = {
  id: SUUID;
  clock: HLC;
  did: Discussion["id"];
  user_id: User["id"];
  text: string;
  media: Media[];
  reply_to?: Message["id"];
  edits: MessageEdits[];
  reactions: Record<User["id"], Record<ReactionStr, SDate>>;
  posted_as_discussion?: Discussion["id"][];
  mentions: Record<User["id"], Mention>;
  flagged_uids?: User["id"][];
  link_previews?: LinkPreviewData[];
  created_at: SDate;
  updated_at: SDate;
  deleted_at?: SDate;
};
// ...
export type Media = ImageMedia | VideoMedia | AudioMedia;
// ...
export type GroupSettings = { 
  mode: "crew" | null; 
  member_mode: "open" | "closed" | "friends_of_friends" 
};

export type Group = {
  id: SULID;
  created_at: SDate;
  created_by: User["id"];
  updated_at: SDate;
  name: string;
  description?: string;
  avatar?: string;
  owner: User["id"];
  admins: User["id"][];
  members: User["id"][];
  joined_at: Record<User["id"], SDate>;
  archived_uids: User["id"][];
  settings: GroupSettings;
  is_public: boolean;
};
// ...
export type Contact = {
  id: SUUID;
  name: string;
  avatar: string | null;
  profile?: UserProfile;
};

export type ContactRequest = {
  id: SUUID;
  from: User["id"];
  state: ContactRequestState;
  created_at: SDate;
};
// ...
export type HydratedContactRequest = ContactRequest & {
  in_common: {
    groups: Group["id"][];
    contacts: Contact["id"][];
  };
};

export type HydratedContact = Contact & {
  contact_request?: ContactRequest["id"];
  invited_by?: Contact["id"];
  in_common: {
    groups: Group["id"][];
    contacts: Contact["id"][];
  };
};

export type HydratedGroup = Group & {
  added_by: Contact["id"];
  in_common: { contacts: Contact["id"][] };
};

export type HydratedDiscussion = Discussion & { messages: Message[] };

export type HydratedInviteLink = InviteLink & {
  contact: Contact;
  in_common: {
    contacts: Contact["id"][];
    groups: Group["id"][];
  };
};

export type NewPostFeedItemPayload = {
  feed_type: "new_post";
  ref_type: "discussion";
  ref: HydratedDiscussion;
};

export type FeedItemPayload =
  | {
    feed_type: "new_request";
    ref_type: "contact_request";
    ref: HydratedContactRequest;
  }
  | { feed_type: "new_friend"; ref_type: "contact"; ref: HydratedContact }
  | {
    feed_type: "new_friend_of_friend";
    ref_type: "contact";
    ref: HydratedContact;
  }
  | {
    feed_type: "new_user_invited_by_friend";
    ref_type: "user";
    ref: HydratedContact;
  }
  | { feed_type: "added_to_group"; ref_type: "group"; ref: HydratedGroup }
  | {
    feed_type: "accepted_invite";
    ref_type: "invite_link";
    ref: HydratedInviteLink;
  }
  | {
    feed_type: "mentioned_in_discussion";
    ref_type: "discussion";
    ref: HydratedDiscussion;
  }
  | NewPostFeedItemPayload;

type BaseFeedItem = {
  id: SUUID;
  created_at: SDate;
  updated_at: SDate;
  uids: User["id"][];
  dismissed_by: User["id"][];
  hidden_for: User["id"][];
  contact: Contact["id"];
  group: Group["id"];
  location_id: Location["id"];
  contact_request: ContactRequest["id"];
  seen_at: Record<User["id"], SDate>;
};

export type ShallowFeedItem = BaseFeedItem & {
  ref_type: FeedItemPayload["ref_type"];
  feed_type: FeedItemPayload["feed_type"];
  ref: string;
};

export type NewPostFeedItem = BaseFeedItem & NewPostFeedItemPayload;
export type FeedItem = BaseFeedItem & FeedItemPayload;
// ...
export type FeedType = "all_posts" | "active_discussions";
export type AllFeedType = "all_posts" | "active_discussions" | "search";

export type SearchQuery = {
  type: "all";
  feedType: "search";
  term: string;
  group_id: undefined;
  contact_id: undefined;
  last_id?: FeedItem["id"];
};

export type MainFeedQuery =
  | {
    type: "all";
    feedType: FeedType;
    group_id: undefined;
    contact_id: undefined;
    location_id: undefined;
    last_id?: FeedItem["id"];
    hidden?: boolean;
  }
  | {
    type: "group";
    group_id?: Group["id"];
    contact_id: undefined;
    location_id: undefined;
    feedType: FeedType;
    last_id?: FeedItem["id"];
    hidden?: boolean;
  }
  | {
    type: "contact";
    contact_id?: User["id"];
    group_id: undefined;
    location_id: undefined;
    feedType: FeedType;
    last_id?: FeedItem["id"];
    hidden?: boolean;
  }
  | {
    type: "location";
    location_id: Location["id"];
    group_id: undefined;
    contact_id: undefined;
    feedType: FeedType;
    last_id?: FeedItem["id"];
    hidden?: boolean;
  };
// ...
export interface LinkPreviewData {
  id: SUUID;
  uri: string;
  created_at: SDate;
  version: number;
  url: string;
  title?: string;
  type: "preview";
  site_name?: string;
  host?: string;
  description?: string;
  media_type?: string;
  images: LinkPreviewDataMedia[];
  videos: LinkPreviewDataMedia[];
  html?: string;
  favicons: string[];
}